requirements.txt
pandas
streamlit
pickle-mixin
numpy
scikit-learn
pandas
streamlit-aggrid
numerize
streamlit_option_menu
streamlit-cards-templates
plotly-express
pygwalker
seaborn
matplotlib
streamlit-extras
SQLAlchemy
psycopg2
pygwalker